hubspot integration added

+++ src/services/apiFunctions/deliveryAreas.js
@@ -14,8 +14,8 @@ class StoreDistrict {
     this.id_store_district = 0
     this.id_district = data.id_district
     this.desc_store_district = data.desc_store_district
-    this.delivery_fee = parseInt(data.delivery_fee)
-    this.value_motoboy = parseInt(data.value_motoboy)
+    this.delivery_fee = parseFloat(data.delivery_fee)
+    this.value_motoboy = parseFloat(data.value_motoboy)
     this.enabled_site_delivery = "Y"
   }
 }
